Move plugin related utils to new file
